requirements.txt
streamlit
google-generativeai
pandas
Pillow
folium
streamlit-folium